Beginning refactoring after finishing up - need to include tiling version
