+++ include/generalUtilities.hpp
@@ -3,6 +3,7 @@
 
 #include <cstdio>
 #include <cstdlib>
+#include <time.h>
 #include <cblas.h>
 #include <cublas_v2.h>
 #include <typeinfo>
