+++ src/Tests/sequentialDecomposerTest.cpp
