+++ src/Utilities/cudaCommunicator.cpp
