+++ src/Wrappers/Slate/slateWrapped.cpp
