--- src/distributedMatrix.cpp
@@ -1,6 +1,0 @@
-#include <distributedMatrix.hpp>
-
-Tile::Tile(int rows, int columns) : rows(rows), columns(columns) 
-{
-
-}
